LayoutTests/imported/w3c/web-platform-tests/event-timing/first-input-shadow-dom-expected.txt


FAIL Event Timing: test first input on shadow DOM. assert_equals: expected Element node <custom-button id="custom_button"></custom-button> but got Element node <button id="inner_button_id">Click me</button>
